refactored version ceck as reaper action
