+++ Plugin/reaper_ultraschall/reaper/Application.cpp
@@ -36,7 +36,6 @@
 #include "SWSVersionCheck.h"
 #include "FileManager.h"
 #include "MessageBox.h"
-#include "AboutAction.h"
 
 namespace ultraschall {
     namespace reaper {
@@ -58,7 +57,7 @@ namespace ultraschall {
         const ServiceStatus Application::Start()
         {
             PRECONDITION_RETURN(HealthCheck(), SERVICE_FAILURE);
-            AboutAction::VersionCheck();
+            reaper::Application& application = reaper::Application::Instance();
           
             return SERVICE_SUCCESS;
         }
