+++ Plugin/reaper_ultraschall/reaper/PluginVersionCheck.cpp
@@ -1,0 +1,48 @@
+////////////////////////////////////////////////////////////////////////////////
+//
+// Copyright (c) 2016 Ultraschall (http://ultraschall.fm)
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+//
+////////////////////////////////////////////////////////////////////////////////
+
+#include <string>
+
+#ifndef WIN32
+#import <Foundation/Foundation.h>
+#import <AppKit/AppKit.h>
+#endif // #ifndef WIN32
+
+#include "PluginVersionCheck.h"
+#include "FileManager.h"
+
+namespace ultraschall {
+  namespace reaper {
+    
+    const std::string QueryPluginVersion()
+    {
+#ifdef WIN32
+      const std::string path = FileManager::ProgramFilesDirectory() + "\\REAPER (x64)\\Plugins\\reaper_ultraschall.dll";
+      return FileManager::ReadVersionFromFile(path);
+#else
+      return "2.2";
+#endif // #ifdef WIN32
+    }
+  }
+}
